refactor(textarea): convert textarea component to typescript
