+++ src/components/textarea/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./textarea.component";
+export type { TextareaProps } from "./textarea.component";
