+++ src/components/textarea/textarea.component.tsx
@@ -1,8 +1,9 @@
 import React, { useRef, useEffect, useContext } from "react";
-import PropTypes from "prop-types";
-import styledSystemPropTypes from "@styled-system/prop-types";
+import { MarginProps } from "styled-system";
 
-import { InputPresentation } from "../../__internal__/input";
+import { IconType } from "../icon";
+import { ValidationProps } from "../../__internal__/validations";
+import { CommonInputProps, InputPresentation } from "../../__internal__/input";
 import FormField from "../../__internal__/form-field";
 import useCharacterCount from "../../hooks/__internal__/useCharacterCount";
 
@@ -18,11 +19,98 @@ import { NewValidationContext } from "../carbon-provider/carbon-provider.compone
 import { ErrorBorder, StyledHintText } from "../textbox/textbox.style";
 import ValidationMessage from "../../__internal__/validation-message";
 
-const marginPropTypes = filterStyledSystemMarginProps(
-  styledSystemPropTypes.space
-);
+// TODO: Change characterLimit type to number - batch with other breaking changes
+export interface TextareaProps
+  extends ValidationProps,
+    MarginProps,
+    Omit<CommonInputProps, "size"> {
+  /** Identifier used for testing purposes, applied to the root element of the component. */
+  "data-component"?: string;
+  /** Identifier used for testing purposes, applied to the root element of the component. */
+  "data-element"?: string;
+  /** Identifier used for testing purposes, applied to the root element of the component. */
+  "data-role"?: string;
+  /** id of the input */
+  id?: string;
+  /** Breakpoint for adaptive label (inline labels change to top aligned). Enables the adaptive behaviour when set */
+  adaptiveLabelBreakpoint?: number;
+  /** Automatically focus the input on component mount */
+  autoFocus?: boolean;
+  /** Character limit of the textarea */
+  characterLimit?: string | number;
+  /** Type of the icon that will be rendered next to the input */
+  children?: React.ReactNode;
+  /** The visible width of the text control, in average character widths */
+  cols?: number;
+  /** If true, the component will be disabled */
+  disabled?: boolean;
+  /** Stop the user typing over the characterLimit */
+  enforceCharacterLimit?: boolean;
+  /** Indicate that error has occurred
+  Pass string to display icon, tooltip and red border
+  Pass true boolean to only display red border */
+  error?: boolean | string;
+  /** Allows the Textareas Height to change based on user input */
+  expandable?: boolean;
+  /** Help content to be displayed under an input */
+  fieldHelp?: React.ReactNode;
+  /** Aria label for rendered help component */
+  helpAriaLabel?: string;
+  /** Indicate additional information
+  Pass string to display icon, tooltip and blue border
+  Pass true boolean to only display blue border */
+  info?: boolean | string;
+  /**
+   * <a href="https://brand.sage.com/d/NdbrveWvNheA/foundations#/icons/icons" target="_blank">List of supported icons</a>
+   *
+   * Icon to display inside of the Textarea
+   */
+  inputIcon?: IconType;
+  /** Width of an input in percentage. Works only when labelInline is true */
+  inputWidth?: number;
+  /** The content of the label for the input */
+  label?: string;
+  /** Inline label alignment */
+  labelAlign?: "left" | "right";
+  /** Text applied to label help tooltip */
+  labelHelp?: React.ReactNode;
+  /** When true, label is placed in line an input */
+  labelInline?: boolean;
+  /** Spacing between label and a field for inline label, given number will be multiplied by base spacing unit (8) */
+  labelSpacing?: 1 | 2;
+  /** Width of a label in percentage. Works only when labelInline is true */
+  labelWidth?: number;
+  /** Name of the input */
+  name?: string;
+  /** Callback fired when the user types in the Textarea */
+  onChange?: (ev: React.ChangeEvent<HTMLInputElement>) => void;
+  /** Placeholder text for the component */
+  placeholder?: string;
+  /** Adds readOnly property */
+  readOnly?: boolean;
+  /** Flag to configure component as mandatory */
+  required?: boolean;
+  /** The number of visible text lines for the control */
+  rows?: number;
+  /** One of type of size to apply to the textarea */
+  size?: "small" | "medium" | "large";
+  /** Message to be displayed in a Tooltip when the user hovers over the help icon */
+  tooltipMessage?: string;
+  /** Overrides the default tooltip position */
+  tooltipPosition?: "top" | "bottom" | "left" | "right";
+  /** When true, validation icon will be placed on label instead of being placed on the input */
+  validationOnLabel?: boolean;
+  /** The value of the Textbox */
+  value?: string;
+  /** Whether to display the character count message in red */
+  warnOverLimit?: boolean;
+  /** Indicate that warning has occurred
+  Pass string to display icon, tooltip and orange border
+  Pass true boolean to only display orange border */
+  warning?: boolean | string;
+}
 
-const Textarea = ({
+export const Textarea = ({
   autoFocus,
   fieldHelp,
   label,
@@ -59,20 +147,21 @@ const Textarea = ({
   "data-role": dataRole,
   helpAriaLabel,
   ...props
-}) => {
+}: TextareaProps) => {
   const { validationRedesignOptIn } = useContext(NewValidationContext);
-  const computeLabelPropValues = (prop) =>
+  const computeLabelPropValues = <T,>(prop: T): undefined | T =>
     validationRedesignOptIn ? undefined : prop;
 
   const { current: id } = useRef(idProp || guid());
 
-  const inputRef = useRef();
+  const inputRef = useRef<HTMLInputElement>(null);
 
   const minHeight = useRef(MIN_HEIGHT);
 
   const expandTextarea = () => {
     const textarea = inputRef.current;
-    if (textarea.scrollHeight > minHeight.current) {
+
+    if (textarea?.scrollHeight && textarea?.scrollHeight > minHeight.current) {
       textarea.style.height = "0px";
       // Set the height so all content is shown
       textarea.style.height = `${Math.max(
@@ -109,7 +198,7 @@ const Textarea = ({
 
   useEffect(() => {
     if (rows) {
-      minHeight.current = inputRef.current.scrollHeight;
+      minHeight.current = inputRef?.current?.scrollHeight || 0;
     }
   }, [rows]);
 
@@ -122,7 +211,7 @@ const Textarea = ({
   useEffect(() => {
     if (expandable) {
       window.addEventListener("resize", expandTextarea);
-      minHeight.current = inputRef.current.clientHeight;
+      minHeight.current = inputRef?.current?.clientHeight || 0;
     }
 
     return () => {
@@ -205,7 +294,6 @@ const Textarea = ({
                 onChange={onChange}
                 disabled={disabled}
                 readOnly={readOnly}
-                labelInline={labelInline}
                 placeholder={disabled ? "" : placeholder}
                 rows={rows}
                 cols={cols}
@@ -238,93 +326,4 @@ const Textarea = ({
   );
 };
 
-Textarea.propTypes = {
-  ...marginPropTypes,
-  /** Automatically focus the input on component mount */
-  autoFocus: PropTypes.bool,
-  /** Identifier used for testing purposes, applied to the root element of the component. */
-  "data-component": PropTypes.string,
-  /** Identifier used for testing purposes, applied to the root element of the component. */
-  "data-element": PropTypes.string,
-  /** Identifier used for testing purposes, applied to the root element of the component. */
-  "data-role": PropTypes.string,
-  /** id of the input */
-  id: PropTypes.string,
-  /** Character limit of the textarea */
-  characterLimit: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-  /** Type of the icon that will be rendered next to the input */
-  children: PropTypes.node,
-  /** The visible width of the text control, in average character widths */
-  cols: PropTypes.number,
-  /** Adds disabled property */
-  disabled: PropTypes.bool,
-  /** Stop the user typing over the characterLimit */
-  enforceCharacterLimit: PropTypes.bool,
-  /** Allows the Textareas Height to change based on user input */
-  expandable: PropTypes.bool,
-  /** Help content to be displayed under an input */
-  fieldHelp: PropTypes.node,
-  /** The content of the label for the input */
-  label: PropTypes.string,
-  /** Text applied to label help tooltip */
-  labelHelp: PropTypes.node,
-  /** Inline label alignment */
-  labelAlign: PropTypes.oneOf(["left", "right"]),
-  /** When true, label is placed in line with an input */
-  labelInline: PropTypes.bool,
-  /** Spacing between label and a field for inline label, given number will be multiplied by base spacing unit (8) */
-  labelSpacing: PropTypes.oneOf([1, 2]),
-  /** Width of a label in percentage. Works only when labelInline is true */
-  labelWidth: PropTypes.number,
-  /** Width of an input in percentage. Works only when labelInline is true */
-  inputWidth: PropTypes.number,
-  /** Name of the input */
-  name: PropTypes.string,
-  /** Callback fired when the user types in the Textarea */
-  onChange: PropTypes.func,
-  /** Placeholder text for the component */
-  placeholder: PropTypes.string,
-  /** Adds readOnly property */
-  readOnly: PropTypes.bool,
-  /** The number of visible text lines for the control */
-  rows: PropTypes.number,
-  /** One of type of size to apply to the textarea */
-  size: PropTypes.oneOf(["small", "medium", "large"]),
-  /** The value of the Textarea */
-  value: PropTypes.string,
-  /** Whether to display the character count message in red */
-  warnOverLimit: PropTypes.bool,
-  /** Indicate that error has occurred
-  Pass string to display icon, tooltip and red border
-  Pass true boolean to only display red border */
-  error: PropTypes.oneOfType([PropTypes.bool, PropTypes.string]),
-  /** Indicate that warning has occurred
-  Pass string to display icon, tooltip and orange border
-  Pass true boolean to only display orange border */
-  warning: PropTypes.oneOfType([PropTypes.bool, PropTypes.string]),
-  /** Indicate additional information
-  Pass string to display icon, tooltip and blue border
-  Pass true boolean to only display blue border */
-  info: PropTypes.oneOfType([PropTypes.bool, PropTypes.string]),
-  /** When true, validation icon will be placed on label instead of being placed on the input */
-  validationOnLabel: PropTypes.bool,
-  /**
-   * <a href="https://brand.sage.com/d/NdbrveWvNheA/foundations#/icons/icons" target="_blank">List of supported icons</a>
-   *
-   * Icon to display inside of the Textarea
-   */
-  inputIcon: PropTypes.string,
-  /** Message to be displayed in a Tooltip when the user hovers over the help icon */
-  tooltipMessage: PropTypes.string,
-  /** Breakpoint for adaptive label (inline labels change to top aligned). Enables the adaptive behaviour when set */
-  adaptiveLabelBreakpoint: PropTypes.number,
-  /** Flag to configure component as mandatory */
-  required: PropTypes.bool,
-  /** Overrides the default tooltip position */
-  tooltipPosition: PropTypes.oneOf(["top", "bottom", "left", "right"]),
-  /** Aria label for rendered help component */
-  helpAriaLabel: PropTypes.string,
-};
-
-export { Textarea as OriginalTextarea };
 export default Textarea;
